+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(desc)
 
-test_check("desc")
+test_check("desc", reporter = "summary")