+++ tests/testthat/test-queries.R
@@ -6,8 +6,28 @@ test_that("get works", {
 
   expect_equal(desc$get("Package"), c(Package = "desc"))
   expect_equal(desc$get("Version"), c(Version = "1.0.0"))
-  expect_equal(desc$get("Author"), c(Author = "Gábor Csárdi"))
+  expect_equal(desc$get("Author"), c(Author = "G\u00e1bor Cs\u00e1rdi"))
   expect_equal(desc$get("Imports"), c(Imports = "\n    R6"))
+})
+
+test_that("get nothing", {
+  desc <- description$new("D1")
+  empty <- structure(character(), names = character())
+  expect_identical(desc$get(character()), empty)
+})
+
+test_that("get_field works", {
+  desc <- description$new("D1")
+
+  expect_identical(desc$get_field("Package"), "desc")
+  expect_identical(desc$get_field("Version"), "1.0.0")
+  expect_identical(desc$get_field("Author"), "G\u00e1bor Cs\u00e1rdi")
+  expect_identical(desc$get_field("Imports"), "\n    R6")
+
+  expect_error(
+    desc$get_field("package"),
+    "Field 'package' not found"
+  )
 })
 
 test_that("get_or_fail works", {
@@ -15,7 +35,7 @@ test_that("get_or_fail works", {
 
   expect_equal(desc$get_or_fail("Package"), c(Package = "desc"))
   expect_equal(desc$get_or_fail("Version"), c(Version = "1.0.0"))
-  expect_equal(desc$get_or_fail("Author"), c(Author = "Gábor Csárdi"))
+  expect_equal(desc$get_or_fail("Author"), c(Author = "G\u00e1bor Cs\u00e1rdi"))
   expect_equal(desc$get_or_fail("Imports"), c(Imports = "\n    R6"))
 
   expect_equal(