+++ inst/NEWS.md
@@ -1,4 +1,21 @@
 
+# devel
+
+* Add `get_field()` method, with easier to use failure and fallback
+  semantics (#62)
+
+* Use the `Encoding` field to read and write DESCRIPTION with the
+  corect encoding. UTF-8 is always used internall by desc. (#52, #53)
+
+* Add `get_built()` function to parse the Built field used in package
+  binaries. (#48, @jimhester)
+
+* `get_deps()` (and `desc_get_deps()`) return a zero-row data frame
+  instead of `NULL` for packages without any dependencies, for consistency.
+
+* Empty `DESCRIPTION` files are handled more gracefully, as are querying
+  no fields with `desc_get()`
+
 # 1.1.1
 
 * Relax the R >= 3.2.0 dependency, R 3.1.0 is enough now.